chore: add formatPhone util
